Remove elasticsearch index name config and infer from code
