+++ api/api/src/main/resources/application.conf
@@ -1,6 +1,5 @@
 es.host=${?es_host}
 es.port=${?es_port}
-es.index.v2=${?es_index_v2}
 es.username=${?es_username}
 es.password=${?es_password}
 es.protocol=${?es_protocol}
