+++ common/elasticsearch/src/test/scala/uk/ac/wellcome/elasticsearch/ElasticsearchIndexCreatorTest.scala
@@ -5,7 +5,6 @@ import scala.concurrent.Future
 import org.scalatest.concurrent.{Eventually, ScalaFutures}
 import org.scalatest.{BeforeAndAfterEach, FunSpec, Matchers}
 import com.sksamuel.elastic4s.ElasticDsl._
-import com.sksamuel.elastic4s.requests.analysis.Analysis
 import com.sksamuel.elastic4s.{RequestFailure, Response}
 import com.sksamuel.elastic4s.requests.indexes.IndexResponse
 import com.sksamuel.elastic4s.requests.searches.SearchResponse
@@ -47,29 +46,29 @@ class ElasticsearchIndexCreatorTest
   )
 
   object TestIndexConfig extends IndexConfig {
-    val mapping = properties(
-      Seq(
-        keywordField("id"),
-        textField("description"),
-        booleanField("visible")
-      ))
-    val analysis = Analysis(Nil)
+    val create = createIndex(createIndexName).mapping(
+      properties(
+        Seq(
+          keywordField("id"),
+          textField("description"),
+          booleanField("visible")
+        )))
   }
 
   object CompatibleTestIndexConfig extends IndexConfig {
-    val mapping = properties(
-      Seq(
-        keywordField("id"),
-        textField("description"),
-        booleanField("visible"),
-        intField("count")
-      ))
-    val analysis = Analysis(Nil)
+    val create = createIndex(createIndexName).mapping(
+      properties(
+        Seq(
+          keywordField("id"),
+          textField("description"),
+          booleanField("visible"),
+          intField("count")
+        )))
   }
 
   it("creates an index into which doc of the expected type can be put") {
     withLocalElasticsearchIndex(TestIndexConfig) { index =>
-      val testObject = TestObject("id", "description", true)
+      val testObject = TestObject("id", "description", visible = true)
       val testObjectJson = toJson(testObject).get
 
       eventually {
@@ -113,44 +112,43 @@ class ElasticsearchIndexCreatorTest
 
   it("updates an already existing index with a compatible mapping") {
     withLocalElasticsearchIndex(TestIndexConfig) { index =>
-      withLocalElasticsearchIndex(CompatibleTestIndexConfig, index = index) {
-        _ =>
-          val compatibleTestObject = CompatibleTestObject(
-            id = "id",
-            description = "description",
-            count = 5,
-            visible = true
-          )
+      withLocalElasticsearchIndex(CompatibleTestIndexConfig) { _ =>
+        val compatibleTestObject = CompatibleTestObject(
+          id = "id",
+          description = "description",
+          count = 5,
+          visible = true
+        )
 
-          val compatibleTestObjectJson = toJson(compatibleTestObject).get
+        val compatibleTestObjectJson = toJson(compatibleTestObject).get
 
-          val futureInsert: Future[Response[IndexResponse]] =
-            elasticClient
-              .execute {
-                indexInto(index.name)
-                  .doc(compatibleTestObjectJson)
-              }
-
-          whenReady(futureInsert) { response =>
-            if (response.isError) { println(response) }
-            response.isError shouldBe false
-
-            eventually {
-              val response: Response[SearchResponse] =
-                elasticClient.execute {
-                  search(index).matchAllQuery()
-                }.await
-
-              val hits = response.result.hits.hits
-
-              hits should have size 1
-
-              assertJsonStringsAreEqual(
-                hits.head.sourceAsString,
-                compatibleTestObjectJson
-              )
+        val futureInsert: Future[Response[IndexResponse]] =
+          elasticClient
+            .execute {
+              indexInto(index.name)
+                .doc(compatibleTestObjectJson)
             }
+
+        whenReady(futureInsert) { response =>
+          if (response.isError) { println(response) }
+          response.isError shouldBe false
+
+          eventually {
+            val response: Response[SearchResponse] =
+              elasticClient.execute {
+                search(index).matchAllQuery()
+              }.await
+
+            val hits = response.result.hits.hits
+
+            hits should have size 1
+
+            assertJsonStringsAreEqual(
+              hits.head.sourceAsString,
+              compatibleTestObjectJson
+            )
           }
+        }
       }
     }
   }
