+++ pipeline/ingestor/src/main/resources/application.conf
@@ -1,6 +1,5 @@
 es.host=${?es_host}
 es.port=${?es_port}
-es.index=${?es_index}
 es.ingest.batchSize=${?es_ingest_batchSize}
 aws.metrics.namespace=${?metrics_namespace}
 es.username=${?es_username}
