+++ pipeline/terraform/stack/service_ingestor.tf
@@ -34,7 +34,6 @@ module "ingestor" {
 
   env_vars = {
     metrics_namespace   = "${local.namespace_hyphen}_ingestor"
-    es_index            = "${var.es_works_index}"
     ingest_queue_id     = "${module.ingestor_queue.id}"
     es_ingest_batchSize = 100
   }
